+++ CHANGES.md
@@ -3,6 +3,7 @@
 - Add `Brr_css.{Highlight,HighlightRegistry,Css}` a binding to the
   CSS Custom Highlight API.
 - Add `Brr.Range` a binding to `Range` objects.
+- Add `Brr.Regexp` a binding to `RegExp` objects.
 - Add `El.text_content` a binding to `Node.textContent`.
 - Add more arguments to `El.scroll_into_view`.
   Thanks to Paul-Elliot Anglès d'Auriac for the patch (#70).